requirements.txt
# Web scraping and parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=5.0.0

# AI/LLM
google-generativeai>=0.8.0

# Document generation
python-docx>=1.1.0
openpyxl>=3.1.0

# PDF extraction
pymupdf>=1.24.0

# Email
# (uses built-in smtplib)

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Date handling
python-dateutil>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

